refactor: button functional to UI file. refactor: clean dead code
